Refactor LineBasketDTO, update BasketService and introduce LineBasketController
- Replace productID and basketID with basketLineID, product, and basket in LineBasketDTO.
- Implement getAllLineBasket method in BasketService to retrieve line basket items by client ID.
- Modify addProductToCart in BasketService to return LineBasketDTO.
- Create LineBasketController to handle API requests for line basket operations, including deletion and quantity updates.
- Introduce LineBasketMapper to ensure object conversion between LineBasket and LineBasketDTO.
- Refactor Basket to return a LineBasket object when adding items.

+++ Ecommerce/src/main/java/com/ilisi/Ecommerce/services/OrderService.java
@@ -27,7 +27,7 @@ public class OrderService {
         return orders.stream().map(orderMapper::toDTO).collect(Collectors.toList());
     }
 
-    public List<OrderDTO> getOrdersByClient(Long clientId) {
+    public List<OrderDTO> getOrdersByClient(int clientId) {
         List<Order> orders = orderRepository.findByClientId(clientId);
         return orders.stream().map(orderMapper::toDTO).collect(Collectors.toList());
     }
@@ -40,7 +40,7 @@ public class OrderService {
     }
 
 
-    public boolean deleteOrder(Long orderId) {
+    public boolean deleteOrder(int orderId) {
         if (!orderRepository.existsById(orderId)) {
             return false;
         }
@@ -49,7 +49,7 @@ public class OrderService {
     }
 
     public OrderDTO getOrderById(int id) {
-        Optional<Order> order = orderRepository.findById((long) id);
+        Optional<Order> order = orderRepository.findById(id);
         if (order.isEmpty()) {
             return null;
         }
@@ -60,7 +60,7 @@ public class OrderService {
     }
 
     public OrderDTO updateOrder(int id, OrderDTO order) {
-        Optional<Order> optionalOrder = orderRepository.findById((long) id);
+        Optional<Order> optionalOrder = orderRepository.findById(id);
 
         if (optionalOrder.isPresent()) {
             Order orderToUpdate = optionalOrder.get();
